Small improvements for USERS controller

+++ server/src/controllers/users.ts
@@ -1,6 +1,7 @@
 import {getConnection} from "typeorm";
 import {User} from "../entity/User";
 import {Request, Response} from "express"
+import {alterKeys} from "../support/support"
 
 const logger = require('../utils/logger')
 const usersRouter = require('express').Router()
@@ -45,16 +46,29 @@ usersRouter.post('/', async (request: Request, response: Response) => {
 usersRouter.put('/:username', async (request: Request, response: Response) => {
   const body = request.body
   const password = await bcrypt.hash(body.password, 10)
+  const username = request.params.username
+
+  const connection = await getConnection();
+  const userRepository = connection.getRepository(User)
+  const usr = await userRepository.findOne({username: username});
+
+  if (usr == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "user not found"
+    })
+  }
+
   await getConnection()
     .createQueryBuilder()
     .update(User)
     .set({
         firstName: body.firstName,
         lastName: body.lastName,
-        username: body.username,
+        // username: body.username,
         password,
     })
-    .where("username = :username", { username: request.params.username })
+    .where("username = :username", { username: username })
     .execute()
     .then(() => {
       return response.status(200).json({
@@ -71,12 +85,23 @@ usersRouter.put('/:username', async (request: Request, response: Response) => {
 })
 
 usersRouter.delete('/:username', async (request: Request, response: Response) => {
-  const body = request.body
+  const username = request.params.username
+
+  const connection = await getConnection();
+  const userRepository = connection.getRepository(User)
+  const usr = await userRepository.findOne({username: username});
+
+  if (usr == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "user not found"
+    })
+  }
   await getConnection()
     .createQueryBuilder()
     .delete()
     .from(User)
-    .where("username = :username", { username: request.params.username })
+    .where("username = :username", { username: username})
     .execute()
     .then(() => {
       return response.status(200).json({
@@ -92,14 +117,16 @@ usersRouter.delete('/:username', async (request: Request, response: Response) =>
     });
 })
 
-usersRouter.get('/names', async (request: Request, response: Response) => {
+usersRouter.get('/usernames', async (request: Request, response: Response) => {
   await getConnection()
     .createQueryBuilder()
     .select("user.username")
     .from(User, "user")
     .execute()
-    .then(item => {
-      return response.status(200).json(item)
+    .then(items => {
+      return response
+      .status(200)
+      .json(alterKeys(items, "user").map((item: { username: string; }) => item.username))
     })
     .catch(error => {
       logger.error(error)
@@ -111,14 +138,27 @@ usersRouter.get('/names', async (request: Request, response: Response) => {
 })
 
 usersRouter.get('/:username', async (request: Request, response: Response) => {
+  const username = request.params.username
+
+  const connection = await getConnection();
+  const userRepository = connection.getRepository(User)
+  const usr = await userRepository.findOne({username: username});
+
+  if (usr == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "user not found"
+    })
+  }
+
   await getConnection()
     .createQueryBuilder()
     .select("user")
     .from(User, "user")
     .where("user.username = :username", {username: request.params.username})
     .execute()
-    .then(item => {
-      return response.status(200).json(item)
+    .then(items => {
+      return response.status(200).json(alterKeys(items, "user"))
     })
     .catch(error => {
       logger.error(error)
@@ -136,7 +176,7 @@ usersRouter.get('/', async (request: Request, response: Response) => {
     .from(User, "user")
     .execute()
     .then(items => {
-      return response.status(200).json(items)
+      return response.status(200).json(alterKeys(items, "user"))
     })
     .catch(error => {
       logger.error(error)
